Updated the object to have a getKeyList along with the getAttributeList since the attribute list is only valid for the attributes added with the addAttribute function.
Updated the example for the object to use the key list instead of the attribute list.

Updated the profiling. Added more profiling for the base object and split it out into multiple pages so that it is faster to run each section.

Added some unit tests for the getAttributeList and getKeyList.