Made the extend private for the base object.
Created the actual code for the filter and option objects.

Created basic unit tests and examples for the filter and option object.

Updated the template unit tests to remove the text comparison on navigation objects since the changes to the template object causes the output html to be unstable. Especially since using the navigation file from the implementation. This may be redone later if we create a stable navigation file that would be used for it.

Reformatted parts of the bookmark unit test.